+++ app/Actions/Course/DeleteCourseAction.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace App\Actions\Course;
 
+use App\Exceptions\ModelHasRelationshipDataException;
 use App\Models\Course;
+use App\Models\Product;
 use Illuminate\Support\Facades\DB;
 
 final readonly class DeleteCourseAction
@@ -15,6 +17,9 @@ final readonly class DeleteCourseAction
     public function handle(Course $course): void
     {
         DB::transaction(function () use ($course): void {
+            if ($course->products()->exists()) {
+                throw new ModelHasRelationshipDataException(Product::class);
+            }
             $course->media()->delete();
             $course->digitalAssets()->detach();
             $course->categories()->detach();
