+++ app/Actions/DigitalAsset/DeleteDigitalAssetAction.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace App\Actions\DigitalAsset;
 
+use App\Exceptions\ModelHasRelationshipDataException;
 use App\Models\DigitalAsset;
+use App\Models\Product;
 use Illuminate\Support\Facades\DB;
 
 final readonly class DeleteDigitalAssetAction
@@ -15,6 +17,9 @@ final readonly class DeleteDigitalAssetAction
     public function handle(DigitalAsset $digitalAsset): void
     {
         DB::transaction(function () use ($digitalAsset): void {
+            if ($digitalAsset->products()->exists()) {
+                throw new ModelHasRelationshipDataException(Product::class);
+            }
             $digitalAsset->media()->delete();
             $digitalAsset->delete();
         });
