+++ app/Actions/Seminar/DeleteSeminarAction.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace App\Actions\Seminar;
 
+use App\Exceptions\ModelHasRelationshipDataException;
+use App\Models\Product;
 use App\Models\Seminar;
 use Illuminate\Support\Facades\DB;
 
@@ -15,6 +17,9 @@ final readonly class DeleteSeminarAction
     public function handle(Seminar $seminar): void
     {
         DB::transaction(function () use ($seminar): void {
+            if ($seminar->products()->exists()) {
+                throw new ModelHasRelationshipDataException(Product::class);
+            }
             $seminar->media()->delete();
             $seminar->digitalAssets()->detach();
             $seminar->categories()->detach();
